Puesto en funcionamiento nuevamente
