+++ .gitignore
@@ -1,2 +1,2 @@
 venv/
-**/_pycache_/
+**/__pycache__/
