+++ control/admin.py
@@ -1,10 +1,10 @@
-from django.contrib import admin
-from .models import *
+# from django.contrib import admin
+# from .models import *
 
-admin.site.register(Company)
-admin.site.register(Branch)
-admin.site.register(User)
-admin.site.register(Product)
-admin.site.register(BranchStock)
-admin.site.register(StockMovement)
+# admin.site.register(Company)
+# admin.site.register(Branch)
+# admin.site.register(User)
+# admin.site.register(Product)
+# admin.site.register(BranchStock)
+# admin.site.register(StockMovement)
