+++ control/models.py
@@ -25,21 +25,21 @@ class Branch(models.Model):
         return self.name
 
 
-# User Model
-class User(models.Model):
-    ROLE_CHOICES = [
-        ('admin', 'Admin'),
-        ('employee', 'Employee'),
-    ]
-    company = models.ForeignKey(Company, on_delete=models.CASCADE)
-    name = models.CharField(max_length=255)
-    email = models.EmailField(unique=True)
-    password = models.CharField(max_length=255)
-    role = models.CharField(max_length=8, choices=ROLE_CHOICES)
-    branch = models.ForeignKey(Branch, null=True, blank=True, on_delete=models.SET_NULL)
+# # User Model
+# class User(models.Model):
+#     ROLE_CHOICES = [
+#         ('admin', 'Admin'),
+#         ('employee', 'Employee'),
+#     ]
+#     company = models.ForeignKey(Company, on_delete=models.CASCADE)
+#     name = models.CharField(max_length=255)
+#     email = models.EmailField(unique=True)
+#     password = models.CharField(max_length=255)
+#     role = models.CharField(max_length=8, choices=ROLE_CHOICES)
+#     branch = models.ForeignKey(Branch, null=True, blank=True, on_delete=models.SET_NULL)
 
-    def __str__(self):
-        return self.name
+#     def __str__(self):
+#         return self.name
 
 
 # Product Model
@@ -73,5 +73,5 @@ class StockMovement(models.Model):
     movement_type = models.CharField(max_length=9, choices=MOVEMENT_TYPE_CHOICES)
     quantity = models.IntegerField()
     date = models.DateTimeField(auto_now_add=True)
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
+    user = models.ForeignKey('user_control.Users', on_delete=models.CASCADE)
     description = models.TextField(null=True, blank=True)
