+++ user_control/admin.py
@@ -1,3 +1,4 @@
 from django.contrib import admin
-
+from user_control.models import Users
 # Register your models here.
+admin.site.register(Users)
