+++ user_control/serializer.py
@@ -1,7 +1,7 @@
 from rest_framework import serializers
-from ..control.models import User
+from .models import Users
 
 class UserSerializer(serializers.ModelSerializer):
     class Meta:
-        model = User
+        model = Users
         fields = '__all__'
