+++ user_control/views.py
@@ -1,9 +1,9 @@
 from django.shortcuts import render
 from rest_framework import viewsets
+from user_control.models import Users
 from .serializer import UserSerializer
-from ..control.models import User
     
 class UserView(viewsets.ModelViewSet):
     serializer_class = UserSerializer
-    queryset = User.objects.all()
+    queryset = Users.objects.all()
